backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["taufiq.ibrahim@gmail.com"]
packages = [
    { include = "src" }
]
keywords = []
license = ""

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = { version = ">=0.115.12,<0.116.0", extras = ["standard"] }
sqlalchemy = ">=2.0.40,<3.0.0"
sqlmodel = ">=0.0.24,<0.0.25"
alembic = ">=1.15.2,<2.0.0"
pydantic-settings = ">=2.8.1,<3.0.0"
passlib = ">=1.7.4,<2.0.0"
pyjwt = "^2.8"
uploadthing-py = "^0.3.3"
bcrypt = "<4.1.0"
boto3 = "^1.37.33"
python-multipart = "^0.0.20"
psycopg2-binary = "^2.9.10"
minio = "^7.2.15"
temporalio = "^1.10.0"
pycryptodome = "^3.22.0"

[tool.poetry.group.test.dependencies]  
pytest-randomly = "^3.16.0"  
pytest = "^8.3.5"
pytest-cov = "^6.1.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest-asyncio = "^0.26.0"
flake8 = "^7.2.0"
isort = "^6.0.1"
flake8-bugbear = "^24.12.12"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
